DB-3650: RCeorganization of tailwind components
